@@ -1,43 +1,40 @@
 source 'http://rubygems.org'
 
-gem 'rails', '3.2.16'
-gem 'sqlite3', '1.3.6'
-gem 'rack', '1.4.5'
-
-group :assets do  
-  gem 'sass-rails', '~> 3.2.3'  
-  gem 'coffee-rails', '~> 3.2.1'  
-  gem 'uglifier', '>= 1.0.3'  
-end
+gem 'rails', '4.1.0'
 
 group :development do
-  gem 'capistrano', '2.15.5'
+  gem 'capistrano', '2.15.5' # Out of date
+  gem 'spring', '~> 1.1.2'
+  gem 'sqlite3', '~> 1.3.9'
 end
 
 group :development, :test do
-  gem 'debugger', '~> 1.6.3'
+  gem 'debugger', '~> 1.6.6'
 end
 
 group :production do
-  gem 'therubyracer', '~> 0.12.0'
+  gem 'therubyracer', '~> 0.12.1'
 end
 
-gem 'bcrypt-ruby', '~> 3.1.2', require: 'bcrypt'
-gem 'bitly', '~> 0.9.0'
-gem 'country_select', '~> 1.2.0'
-gem 'jquery-rails', '~> 3.0.4'
-gem 'kaminari', '~> 0.15.0'
-gem 'kramdown', '~> 1.3.0'
-gem 'local_time', '~> 0.2.0'
+gem 'bcrypt', '~> 3.1.7'
+gem 'bitly', '~> 0.10.1'
+gem 'coffee-rails', '~> 4.0.1'
+gem 'country_select', '~> 1.3.1'
+gem 'jquery-rails', '~> 3.1.0'
+gem 'kaminari', '~> 0.15.1'
+gem 'kramdown', '~> 1.3.3'
+gem 'local_time', '~> 0.3.0'
 gem 'mini_magick', '~> 3.7.0'
-gem 'nokogiri', '~> 1.6.0'
+gem 'nokogiri', '~> 1.6.1'
 gem 'nested_form', '~> 0.3.2'
-gem 'paperclip', '3.4.0'
-gem 'pg', '~> 0.17.0'
-gem 'pundit', '~> 0.2.2'
+gem 'paperclip', '~> 4.1.1'
+gem 'pg', '~> 0.17.1'
+gem 'pundit', '~> 0.2.3'
 gem 'rails_autolink', '~> 1.1.5'
-gem 'recaptcha', '~> 0.3.5', require: 'recaptcha/rails'
-gem 'sanitize', '~> 2.0.6'
-gem 'stringex', '~> 2.1.2'
-gem 'strong_parameters', '~> 0.2.3'
-gem 'twitter', '~> 4.8.1'
+gem 'recaptcha', '~> 0.3.6', require: 'recaptcha/rails'
+gem 'sanitize', '~> 2.1.0'
+gem 'sass-rails', '~> 4.0.3'
+gem 'stringex', '~> 2.5.2'
+gem 'turbolinks', '~> 2.2.2'
+gem 'twitter', '~> 4.8.1' # Out of date
+gem 'uglifier', '>= 2.5.0'
