@@ -1,19 +1,15 @@
 class MusicTracksController < ApplicationController
 
-  before_filter :categories
+  before_action :load_categories, except: [:show, :destroy]
+  before_action :load_music_track, except: [:index, :new, :create]
   
   def index
     if params[:dragoon_id]
-      raise "Dragoon not found." unless @dragoon = Dragoon.find_by_url(params[:dragoon_id])
+      raise "Dragoon not found." unless @dragoon = Dragoon.find_by(url: params[:dragoon_id])
       @music_tracks = policy_scope(MusicTrack.joins(:contributions).where(contributions: { dragoon_id: @dragoon.id }).order(:name).page(params[:page]))
     else
       @music_tracks = policy_scope(MusicTrack.order(:name).page(params[:page]))
     end
-  end
-
-  def show
-    @music_track = MusicTrack.find_by_url(params[:id])
-    authorize @music_track
   end
 
   def new
@@ -31,14 +27,7 @@ class MusicTracksController < ApplicationController
     end
   end
 
-  def edit
-    @music_track = MusicTrack.find_by_url(params[:id])
-    authorize @music_track
-  end
-  
   def update
-    @music_track = MusicTrack.find_by_url(music_track_params)
-    authorize @music_track
     params[:music_track][:dragoon_ids] ||= []
     if @music_track.update_attributes(params[:music_track])
       redirect_to @music_track, notice: "Successfully updated music track."
@@ -48,8 +37,6 @@ class MusicTracksController < ApplicationController
   end
 
   def destroy
-    @music_track = MusicTrack.find_by_url(params[:id])
-    authorize @music_track
     @music_track.destroy
     redirect_to music_tracks_path, notice: "Successfully destroyed music track."
   end
@@ -72,8 +59,13 @@ class MusicTracksController < ApplicationController
     )
   end
 
-  def categories
+  def load_categories
     @categories = CategoryPolicy::Scope.new(current_user, Category.where(category_type: :music_track).order(:name)).resolve
+  end
+
+  def load_music_track
+    @music_track = MusicTrack.find_by url: params[:id]
+    authorize @music_track
   end
 
 end
