@@ -1,19 +1,15 @@
 class PicturesController < ApplicationController
 
-  before_filter :categories
+  before_action :load_categories, except: [:show, :destroy]
+  before_action :load_picture, except: [:index, :new, :create]
   
   def index
     if params[:dragoon_id]
-      raise "Dragoon not found." unless @dragoon = Dragoon.find_by_url(params[:dragoon_id])
+      raise "Dragoon not found." unless @dragoon = Dragoon.find_by(url: params[:dragoon_id])
       @pictures = policy_scope(Picture.joins(:contributions).where(contributions: { dragoon_id: @dragoon.id }).order(:name).page(params[:page]))
     else
       @pictures = policy_scope(Picture.order(:name).page(params[:page]))
     end
-  end
-
-  def show
-    @picture = Picture.find_by_url(params[:id])
-    authorize @picture
   end
 
   def new
@@ -30,15 +26,8 @@ class PicturesController < ApplicationController
       render :new
     end
   end
-
-  def edit
-    @picture = Picture.find_by_url(params[:id])
-    authorize @picture
-  end
   
   def update
-    @picture = Picture.find_by_url(params[:id])
-    authorize @picture
     params[:picture][:dragoon_ids] ||= []
     if @picture.update_attributes(picture_params)
       redirect_to @picture, notice: "Successfully updated picture."
@@ -48,8 +37,6 @@ class PicturesController < ApplicationController
   end
 
   def destroy
-    @picture = Picture.find_by_url(params[:id])
-    authorize @picture
     @picture.destroy
     redirect_to pictures_path, notice: "Successfully destroyed picture."
   end
@@ -69,8 +56,13 @@ class PicturesController < ApplicationController
     )
   end
 
-  def categories
+  def load_categories
     @categories = CategoryPolicy::Scope.new(current_user, Category.where(category_type: :picture).order(:name)).resolve
+  end
+
+  def load_picture
+    @picture = Picture.find_by url: params[:id]
+    authorize @picture
   end
 
 end
