@@ -1,8 +1,10 @@
 class QuizzesController < ApplicationController
   
+  before_action :load_quiz, except: [:index, :new, :create]
+
   def index
     if params[:dragoon_id]
-      raise "Dragoon not found." unless @dragoon = Dragoon.find_by_url(params[:dragoon_id])
+      raise "Dragoon not found." unless @dragoon = Dragoon.find_by(url: params[:dragoon_id])
       @quizzes = policy_scope(Quiz.joins(:contributions).where(contributions: { dragoon_id: @dragoon.id }).order(:name).page(params[:page]))
     else
       @quizzes = policy_scope(Quiz.order(:name).page(params[:page]))
@@ -10,8 +12,6 @@ class QuizzesController < ApplicationController
   end
 
   def show
-    @quiz = Quiz.find_by_url(params[:id])
-    authorize @quiz
     if params[:commit]
       if params[:results]
         if params[:results].count < @quiz.quiz_questions.count
@@ -43,17 +43,13 @@ class QuizzesController < ApplicationController
   end
 
   def edit
-    @quiz = Quiz.find_by_url(params[:id])
-    authorize @quiz
     quiz_question = @quiz.quiz_questions.build  
     3.times { quiz_question.quiz_answers.build }  
   end
   
   def update
-    @quiz = Quiz.find_by_url(quiz_params)
-    authorize @quiz
     params[:quiz][:dragoon_ids] ||= []  
-    if @quiz.update_attributes(params[:quiz])
+    if @quiz.update_attributes(quiz_params)
       redirect_to @quiz, notice: "Successfully updated quiz."
     else
       render :edit
@@ -61,8 +57,6 @@ class QuizzesController < ApplicationController
   end
 
   def destroy
-    @quiz = Quiz.find_by_url(params[:id])
-    authorize @quiz
     @quiz.destroy
     redirect_to quizzes_path, notice: "Successfully destroyed quiz."
   end
@@ -78,6 +72,11 @@ class QuizzesController < ApplicationController
       dragoon_ids: [],
       encyclopaedia_entry_ids: []
     )
+  end
+
+  def load_quiz
+    @quiz = Quiz.find_by url: params[:id]
+    authorize @quiz
   end
 
 end
