@@ -1,20 +1,18 @@
 class StoriesController < ApplicationController
 
-  before_filter :categories
-  
+  before_action :load_categories, except: [:show, :destroy]
+  before_action :load_story, except: [:index, :new, :create]
+    
   def index
     if params[:dragoon_id]
-      raise "Dragoon not found." unless @dragoon = Dragoon.find_by_url(params[:dragoon_id])
+      raise "Dragoon not found." unless @dragoon = Dragoon.find_by(url: params[:dragoon_id])
       @stories = policy_scope(Story.joins(:contributions).where(contributions: { dragoon_id: @dragoon.id }).order(:name).page(params[:page]))
     else
       @stories = policy_scope(Story.order(:name).page(params[:page]))
     end
   end
 
-  def show
-    @story = Story.find_by_url(params[:id])
-    authorize @story
-    
+  def show    
     @chapter_count = ChapterPolicy::Scope.new(current_user, @story.chapters).resolve.count
     @prologues = ChapterPolicy::Scope.new(current_user, @story.chapters.where(chapter_type: :prologue).order(:number)).resolve
     @regular_chapters = ChapterPolicy::Scope.new(current_user, @story.chapters.where(chapter_type: :regular_chapter).order(:number)).resolve
@@ -36,25 +34,17 @@ class StoriesController < ApplicationController
     end
   end
   
-  def edit
-    @story = Story.find_by_url(params[:id])
-    authorize @story
-  end
-
   def update
-    @story = Story.find_by_url(story_params)
-    authorize @story
     params[:story][:dragoon_ids] ||= []
-    if @story.update_attributes(params[:story])      
+    if @story.update_attributes(story_params)      
       redirect_to @story, notice: "Successfully updated story."
     else
       render :edit
     end
   end
 
-  def destroy    
+  def destroy
     @story.destroy
-    authorize @story
     redirect_to stories_path, notice: "Successfully destroyed story."
   end
   
@@ -73,8 +63,13 @@ class StoriesController < ApplicationController
     )
   end
 
-  def categories
+  def load_categories
     @categories = CategoryPolicy::Scope.new(current_user, Category.where(category_type: :story).order(:name)).resolve
+  end
+
+  def load_story
+    @story = Story.find_by url: params[:id]
+    authorize @story
   end
 
 end
