@@ -1,19 +1,15 @@
 class VideosController < ApplicationController
 
-  before_filter :categories
+  before_action :load_categories, expect: [:show, :destroy]
+  before_action :load_video, except: [:index, :new, :create]
   
   def index
     if params[:dragoon_id]
-      raise "Dragoon not found." unless @dragoon = Dragoon.find_by_url(params[:dragoon_id])
+      raise "Dragoon not found." unless @dragoon = Dragoon.find_by(url: params[:dragoon_id])
       @videos = policy_scope(Video.joins(:contributions).where(contributions: { dragoon_id: @dragoon.id }).order(:name).page(params[:page]))
     else
       @videos = policy_scope(Video.order(:name).page(params[:page]))
     end
-  end
-
-  def show
-    @video = Video.find_by_url(params[:id])
-    authorize @video
   end
 
   def new
@@ -31,14 +27,7 @@ class VideosController < ApplicationController
     end
   end
 
-  def edit
-    @video = Video.find_by_url(params[:id])
-    authorize @video
-  end
-  
   def update
-    @video = Video.find_by_url(params[:id])
-    authorize @video
     params[:video][:dragoon_ids] ||= []
     if @video.update_attributes(video_params)
       redirect_to @video, notice: "Successfully updated video."
@@ -48,8 +37,6 @@ class VideosController < ApplicationController
   end
 
   def destroy
-    @video = Video.find_by_url(params[:id])
-    authorize @video
     @video.destroy
     redirect_to videos_path, notice: "Successfully destroyed video."
   end
@@ -71,8 +58,13 @@ class VideosController < ApplicationController
     )
   end
 
-  def categories
+  def load_categories
     @categories = CategoryPolicy::Scope.new(current_user, Category.where(category_type: :video).order(:name)).resolve
+  end
+
+  def load_video
+    @video = Video.find_by url: params[:id]
+    authorize @video
   end
 
 end
