@@ -1,11 +1,14 @@
 class ApplicationController < ActionController::Base
   
+  # Prevent CSRF attacks by raising an exception.
+  # For APIs, you may want to use :null_session instead.
+  protect_from_forgery with: :exception
+
   include Pundit
   rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
   after_filter :verify_authorized, except: :index
   after_filter :verify_policy_scoped, only: :index
 
-  protect_from_forgery
   helper_method :current_user
   helper_method :current_dragoon
   helper_method :current_ability
