@@ -43,7 +43,7 @@ namespace :deploy do
   namespace :assets do
     desc "Precompile assets on local machine and upload them to the server."
     task :precompile, roles: :web, except: {no_release: true} do
-      run_locally "bundle exec rake assets:precompile RAILS_ENV=development"
+      run_locally "bundle exec rake assets:precompile RAILS_ENV=production"
       find_servers_for_task(current_task).each do |server|
         run_locally "rsync -vr --exclude='.DS_Store' public/assets #{user}@#{server.host}:#{shared_path}/"
       end
